+++ src/main/java/kr/co/conceptbe/branch/domain/persistense/BranchRepository.java
@@ -1,7 +1,7 @@
-package kr.co.conceptbe.common.entity.domain.persistence;
+package kr.co.conceptbe.branch.domain.persistense;
 
 import java.util.List;
-import kr.co.conceptbe.common.entity.domain.Branch;
+import kr.co.conceptbe.branch.domain.Branch;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface BranchRepository extends JpaRepository<Branch, Long> {
