+++ src/main/java/kr/co/conceptbe/idea/presentation/dto/response/IdeaResponse.java
@@ -2,7 +2,7 @@ package kr.co.conceptbe.idea.presentation.dto.response;
 
 import java.time.LocalDateTime;
 import java.util.List;
-import kr.co.conceptbe.common.entity.domain.Branch;
+import kr.co.conceptbe.branch.domain.Branch;
 import kr.co.conceptbe.common.entity.domain.TeamRecruitment;
 import kr.co.conceptbe.idea.domain.Idea;
 import kr.co.conceptbe.idea.domain.IdeaBranch;
